front: global devient un controller; back: extériorisation des données database et clé de token, suppression des images pour github

+++ backend/controllers/articles.js
@@ -1,3 +1,4 @@
+const User = require('../models/User');
 const Article = require('../models/Article');
 const Comment = require('../models/Comment');
 const fs = require('fs');
@@ -60,24 +61,27 @@ exports.modifyArticle = (req, res, next) => {
         if (!article) {
             return res.status(404).json({ error: 'Article non trouvé !' });
         } else {
-            if (article.authorId == req.body.userId) { // on vérifie que l'auteur connu de l'article soit bien l'initiateur de la requête
-                if (article.imageUrl || req.file) { // si l'article possède une image, on  la supprime
-                    fs.unlink(`${article.imageUrl}`, () => {});
-                }
-                const articleObject = req.file ? {
-                    ...req.body,
-                    imageUrl: `images/${req.file.filename}`
-                } : {...req.body };
-                Article.updateOne(articleObject, function(result) {
-                    if (!result) {
-                        return res.status(201).json({ message: 'Article modifié !' })
-                    } else {
-                        return res.status(400).json({ error: 'Une erreur est survenue !' });
+            User.findOne("id", req.body.userId, function(user) {
+                if (user.admin != 1 && article.authorId != req.body.userId) {
+                    return res.status(400).json({ error: "Vous n'êtes pas l'auteur !" });
+                } else {
+                    if (article.imageUrl || req.file) { // si l'article possède une image, on  la supprime
+                        fs.unlink(`${article.imageUrl}`, () => {});
                     }
-                });
-            } else {
-                return res.status(400).json({ error: 'Bad request !' });
-            }
+                    const articleObject = req.file ? {
+                        ...req.body,
+                        imageUrl: `images/${req.file.filename}`
+                    } : {...req.body };
+                    Article.updateOne(articleObject, function(result) {
+                        if (!result) {
+                            return res.status(201).json({ message: 'Article modifié !' })
+                        } else {
+                            return res.status(400).json({ error: 'Une erreur est survenue !' });
+                        }
+                    });
+
+                }
+            });
         }
     });
 }
@@ -87,26 +91,28 @@ exports.deleteArticle = (req, res, next) => {
         if (!article) {
             return res.status(404).json({ error });
         } else {
-            if (article.authorId != req.body.userId) { // on vérifie que l'auteur connu de l'article soit bien l'initiateur de la requête
-                return res.status(400).json({ error: 'Bad request !' });
-            } else {
-                Comment.deleteByArticleId(req.params.id, function(result) {
-                    if (result) {
-                        return res.status(400).json({ error: 'Une erreur est survenue lors de la suppression des commentaires !' });
+            User.findOne("id", req.body.userId, function(user) {
+                if (user.admin != 1 && article.authorId != req.body.userId) {
+                    return res.status(400).json({ error: "Vous n'êtes pas l'auteur !" });
+                } else {
+                    Comment.deleteByArticleId(req.params.id, function(result) {
+                        if (result) {
+                            return res.status(400).json({ error: 'Une erreur est survenue lors de la suppression des commentaires !' });
+                        }
+                    });
+                    if (article.imageUrl) { // s'il y a une image, on la supprime
+                        fs.unlink(`${article.imageUrl}`, () => {});
                     }
-                });
-                if (article.imageUrl) { // s'il y a une image, on la supprime
-                    fs.unlink(`${article.imageUrl}`, () => {});
-                }
-                Article.deleteOne(req.params.id, function(result) {
-                    if (!result) {
-                        res.status(200).json({ message: "Article supprimé !" });
-                    } else {
-                        return res.status(400).json({ error: "Une erreur est survenue lors de la suppression de l'article !" });
-                    }
-                });
+                    Article.deleteOne(req.params.id, function(result) {
+                        if (!result) {
+                            res.status(200).json({ message: "Article supprimé !" });
+                        } else {
+                            return res.status(400).json({ error: "Une erreur est survenue lors de la suppression de l'article !" });
+                        }
+                    });
 
-            }
+                }
+            });
         }
     });
 }
